Better Shutdown Process for Pipes
Handle all of the edge cases when shutting down in Pipes mode.
e.g.
If the pipes are broken (tried to write after the program died)
If the fifos don't exist anymore (sometimes tempfile cleans them up
before the destructor finishes when certain errors happen)
If the object attributes for the streams and fifo paths aren't set (this
can happen if the constructor didn't finish. e.g. the user cancels while
waiting on a connection)

+++ sploit/comm.py
@@ -132,8 +132,11 @@ class Pipes:
         print("Connected!")
 
     def __del__(self):
-        self.stdout.close()
-        self.stdin.close()
-        os.unlink(self.pathin)
-        os.unlink(self.pathout)
+        try:
+            if getattr(self,'stdout',None) : self.stdout.close()
+            if getattr(self,'stdin',None) : self.stdin.close()
+        except BrokenPipeError:
+            pass
+        if getattr(self,'pathin',None) and os.path.exists(self.pathin) : os.unlink(self.pathin)
+        if getattr(self,'pathout',None) and os.path.exists(self.pathout) : os.unlink(self.pathout)
